# backend/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:8080
info:
  contact:
    email: dev@albyeicorp.com
    name: Albyei Corp Engineering
  description: API untuk sistem rekrutmen Wowrack/Albyei Corp. Mendukung multi-role
    (Admin, HR, Hiring Manager, Candidate).
  termsOfService: http://swagger.io/terms/
  title: Wowrack Recruitment API
  version: "1.0"
paths:
...
  /hr/news-cultures:
    get:
      description: Get paginated list of news articles with optional filters (HR only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by published status
        enum:
        - "true"
        - "false"
        in: query
        name: published
        type: string
      - description: Search by title
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get all news/culture articles
      tags:
      - News
    post:
      consumes:
      - multipart/form-data
      description: Create a new news or culture article with image and optional gallery
        (HR only)
      parameters:
      - description: Article title (min 5, max 200)
        in: formData
        name: title
        required: true
        type: string
      - description: Article content (min 20)
        in: formData
        name: content
        required: true
        type: string
      - description: Short excerpt (max 500)
        in: formData
        name: excerpt
        required: true
        type: string
      - description: Publish immediately
        in: formData
        name: published
        type: boolean
      - description: Main image
        in: formData
        name: image
        required: true
        type: file
      - description: Gallery images (multiple)
        in: formData
        name: gallery
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create news/culture article
      tags:
      - News
  /hr/news-cultures/{id}:
    delete:
      description: Permanently delete a news article and all associated photos (HR
        only)
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete news/culture article
      tags:
      - News
    get:
      description: Get a single news/culture article by its ID (HR only)
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get news by ID
      tags:
      - News
    put:
      consumes:
      - multipart/form-data
      description: Update an existing news article with optional new image/gallery
        (HR only)
      parameters:
      - description: News ID
        in: path
        name: id
        required: true
        type: integer
      - description: Article title
        in: formData
        name: title
        type: string
      - description: Article content
        in: formData
        name: content
        type: string
      - description: Excerpt
        in: formData
        name: excerpt
        type: string
      - description: Published status
        in: formData
        name: published
        type: boolean
      - description: New main image
        in: formData
        name: image
        type: file
      - description: New gallery images
        in: formData
        name: gallery
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update news/culture article
      tags:
      - News
...
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"